fix: Move images folder to tthe asset folder and ddone the changes in the files accordingly
