+++ app/components/AuthLayout.tsx
@@ -1,4 +1,4 @@
-import backgroundImage from '../images/background-auth.jpg'
+import backgroundImage from '../../assets/images/background-auth.jpg'
 
 export function AuthLayout({ children }: any) {
   return (
