+++ app/components/CallToAction.tsx
@@ -1,5 +1,5 @@
 // import { Container } from '../components/Container'
-import backgroundImage from '../images/background-call-to-action.jpg'
+import backgroundImage from '../../assets/images/background-call-to-action.jpg'
 
 export function CallToAction() {
   return (
