+++ app/routes/successlogin/index.tsx
@@ -9,7 +9,7 @@ import {
   FormikCheckbox,
   FormikInput,
 } from '../../components/Common/FormikInput'
-import logo from '../../images/logos/quicklook-icon.svg'
+import logo from '../../../assets/images/logos/quicklook-icon.svg'
 import { Form } from '@remix-run/react'
 
 export const action: ActionFunction = async ({ request }) => {
